Add loader class for helper methods used in all classes and have each class inherit instead of repeating code
