+++ specs/spec_helper.rb
@@ -11,3 +11,4 @@ require_relative '../lib/room'
 require_relative '../lib/reservation'
 require_relative '../lib/booking'
 require_relative '../lib/block'
+require_relative '../lib/loader'
